# Conflicts:
#	src/main/java/balancetalk/module/post/dto/BalanceOptionRequest.java

+++ src/main/java/balancetalk/module/member/application/MemberService.java
@@ -4,9 +4,12 @@ import balancetalk.global.exception.BalanceTalkException;
 import balancetalk.global.exception.ErrorCode;
 import balancetalk.global.jwt.JwtTokenProvider;
 import balancetalk.global.redis.application.RedisService;
+import balancetalk.module.file.domain.File;
+import balancetalk.module.file.domain.FileRepository;
 import balancetalk.module.member.domain.Member;
 import balancetalk.module.member.domain.MemberRepository;
 import balancetalk.module.member.dto.*;
+import io.jsonwebtoken.ExpiredJwtException;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -22,6 +25,8 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static balancetalk.global.exception.ErrorCode.*;
+
 @Slf4j
 @Service
 @RequiredArgsConstructor
@@ -30,21 +35,29 @@ public class MemberService {
     private final JwtTokenProvider jwtTokenProvider;
     private final AuthenticationManager authenticationManager;
     private final MemberRepository memberRepository;
+    private final FileRepository fileRepository;
     private final PasswordEncoder passwordEncoder;
     private final RedisService redisService;
 
     @Transactional
     public Long join(final JoinRequest joinRequest) {
-        joinRequest.setPassword(passwordEncoder.encode(joinRequest.getPassword()));
-        if (memberRepository.existsByNickname(joinRequest.getNickname())) {
-            throw new BalanceTalkException(ErrorCode.ALREADY_REGISTERED_NICKNAME);
-        }
         if (memberRepository.existsByEmail(joinRequest.getEmail())) {
-            throw new BalanceTalkException(ErrorCode.ALREADY_REGISTERED_EMAIL);
+            throw new BalanceTalkException(ALREADY_REGISTERED_EMAIL);
         }
-        Member member = joinRequest.toEntity();
+        if (memberRepository.existsByNickname(joinRequest.getNickname())) {
+            throw new BalanceTalkException(ALREADY_REGISTERED_NICKNAME);
+        }
+        joinRequest.setPassword(passwordEncoder.encode(joinRequest.getPassword()));
+        File profilePhoto = null;
+        if (joinRequest.getProfilePhoto() != null && !joinRequest.getProfilePhoto().isEmpty()) {
+            profilePhoto = fileRepository.findByStoredName(joinRequest.getProfilePhoto())
+                    .orElseThrow(() -> new BalanceTalkException(NOT_FOUND_FILE));
+        }
+
+        Member member = joinRequest.toEntity(profilePhoto);
         return memberRepository.save(member).getId();
     }
+
 
     @Transactional
     public TokenDto login(final LoginRequest loginRequest) {
@@ -57,8 +70,8 @@ public class MemberService {
         Authentication authentication = authenticationManager.authenticate(
                 new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword())
         );
-        String refreshToken = jwtTokenProvider.createRefreshToken(authentication);
-        return jwtTokenProvider.reissueToken(refreshToken); // 만료되었다면, 재발급
+        String refreshToken = jwtTokenProvider.createRefreshToken(authentication, member.getId());
+        return jwtTokenProvider.reissueToken(refreshToken, member.getId()); // 만료되었다면, 재발급
     }
 
     @Transactional(readOnly = true)
@@ -110,6 +123,9 @@ public class MemberService {
     @Transactional
     public void logout(){
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication == null) {
+            throw new BalanceTalkException(ErrorCode.AUTHENTICATION_REQUIRED);
+        }
         String username = authentication.getName();
         redisService.deleteValues(username);
     }
